refactor(test): P5-1 follow-up — extract yoga logical-edge + A4 OOS
Continue test_widget_bridge.cpp split. Extracts the 497-line bundle
of pulp #1542 (yoga logical-edge fan-out) + pulp #1434 A4 Bundles 2-7
(css NOT-IMPL closure) into a sibling TU.

Coverage:
  * #1542: marginInline / marginBlock / paddingInline / paddingBlock
    + inset shorthand through CSS translator into Yoga's logical-edge
    slots (LTR/RTL-aware writing-mode-respecting layout).
  * #1434 A4 Bundles 2-7: catalog hygiene tests for CSS properties
    Pulp deliberately treats as no-ops or sensible fallbacks.

* test_widget_bridge.cpp: 6,711 → 6,214 lines (-497)
* New target pulp-test-widget-bridge-yoga-a4-oos: 537 lines,
  14 TEST_CASEs, 70 assertions.

This is commit 1 of 2 in a bundled-cuts PR (Wave 2 cheap-wiring
follows; single version bump at the end).
